refector:Exceções inseridas nas classes de serviço

+++ src/api/services/user.service.js
@@ -1,8 +1,15 @@
 const UserModel = require("../models/schemas/user.model");
+const BadRequestException = require("../models/exceptions/bad-request.exception");
+const NotFoundRequestException = require("../models/exceptions/not-found-request.exception");
 
 class UserService {
 
     async create(data) {
+        const findByEmail = await this.readByEmail(data.email);
+        if (findByEmail) {
+            throw new BadRequestException("E-mail já cadastrado");
+        }
+
         return await UserModel.create(data);
     }
 
@@ -18,7 +25,12 @@ class UserService {
     }
 
     async readById(id) {
-        return await UserModel.findById(id);
+        const findById = await UserModel.findById(id);
+        if (!findById) {
+            throw new NotFoundRequestException("Usuário não encontrado");
+        }
+
+        return findById;
     }
 
     async readByEmailAndPassword(email, password) {
@@ -30,11 +42,28 @@ class UserService {
     }
 
     async update(id, data) {
+        const findById = await this.readById(id);
+        if (!findById) {
+            throw new NotFoundRequestException("Usuário não encontrado");
+        }
+
+        if (findById.email != data?.email) {
+            const findByEmail = await this.readByEmail(data.email);
+            if (findByEmail && findByEmail._id != findById._id) {
+                throw new BadRequestException("E-mail já cadastrado");
+            }
+        }
+
         return await UserModel.findByIdAndUpdate(id, data, { new: true });
     }
 
     async delete(id) {
-        return await UserModel.findByIdAndRemove(id);
+        const findById = await UserModel.findById(id);
+        if (!findById) {
+            throw new NotFoundRequestException("Usuário não encontrado");
+        }
+
+        await UserModel.findByIdAndRemove(id);
     }
 }
